~ WIP fixing Matrix encoding
