+++ opc-ua-stack/encoding-xml/src/test/java/org/eclipse/milo/opcua/stack/core/encoding/xml/OpcUaXmlEncoderTest.java
@@ -12,18 +12,11 @@ package org.eclipse.milo.opcua.stack.core.encoding.xml;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.UUID;
 import org.eclipse.milo.opcua.stack.core.UaSerializationException;
 import org.eclipse.milo.opcua.stack.core.encoding.DefaultEncodingContext;
 import org.eclipse.milo.opcua.stack.core.encoding.EncodingContext;
-import org.eclipse.milo.opcua.stack.core.types.UaStructuredType;
 import org.eclipse.milo.opcua.stack.core.types.builtin.*;
-import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UByte;
-import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UInteger;
-import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.ULong;
-import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UShort;
 import org.jspecify.annotations.Nullable;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -51,945 +44,20 @@ class OpcUaXmlEncoderTest {
     }
   }
 
-  @Nested
-  class ScalarTests {
+  @ParameterizedTest(name = "matrix = {0}")
+  @MethodSource(
+      "org.eclipse.milo.opcua.stack.core.encoding.xml.args.MatrixArguments#matrixOfBuiltinTypeArguments")
+  void encodeMatrixOfBuiltinType(@Nullable Matrix matrix, String expected) {
+    var encoder = new OpcUaXmlEncoder(context);
+    encoder.encodeMatrix("Test", matrix);
 
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#booleanArguments")
-    void encodeBoolean(Boolean value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeBoolean("Test", value);
+    String actual = encoder.getDocumentXml();
 
-      String actual = encoder.getDocumentXml();
+    Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
 
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
+    maybePrintXml(diff, expected, actual);
 
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#sByteArguments")
-    void encodeSByte(Byte value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeSByte("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#byteArguments")
-    void encodeByte(UByte value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeByte("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#int16Arguments")
-    void encodeInt16(Short value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt16("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#uInt16Arguments")
-    void encodeUInt16(UShort value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt16("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#int32Arguments")
-    void encodeInt32(Integer value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt32("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#uInt32Arguments")
-    void encodeUInt32(UInteger value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt32("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#int64Arguments")
-    void encodeInt64(Long value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt64("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#uInt64Arguments")
-    void encodeUInt64(ULong value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt64("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#floatArguments")
-    void encodeFloat(Float value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeFloat("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#doubleArguments")
-    void encodeDouble(Double value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDouble("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "value = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#stringArguments")
-    void encodeString(@Nullable String value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeString("Test", value);
-
-      String actual = encoder.getDocumentXml();
-
-      if (value == null) {
-        // When encoding a null string the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "dateTime = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#dateTimeArguments")
-    void encodeDateTime(@Nullable DateTime dateTime, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDateTime("Test", dateTime);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "guid = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#guidArguments")
-    void encodeGuid(@Nullable UUID guid, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeGuid("Test", guid);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "byteString = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#byteStringArguments")
-    void encodeByteString(@Nullable ByteString byteString, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeByteString("Test", byteString);
-
-      String actual = encoder.getDocumentXml();
-
-      if (byteString == null) {
-        // When encoding a null ByteString the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "xmlElement = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#xmlElementArguments")
-    void encodeXmlElement(@Nullable XmlElement xmlElement, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeXmlElement("Test", xmlElement);
-
-      String actual = encoder.getDocumentXml();
-
-      if (xmlElement == null) {
-        // When encoding a null XmlElement the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "nodeId = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#nodeIdArguments")
-    void encodeNodeId(NodeId nodeId, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeNodeId("Test", nodeId);
-
-      String actual = encoder.getDocumentXml();
-
-      if (nodeId == null) {
-        // When encoding a null NodeId the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "expandedNodeId = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#expandedNodeIdArguments")
-    void encodeExpandedNodeId(@Nullable ExpandedNodeId expandedNodeId, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeExpandedNodeId("Test", expandedNodeId);
-
-      String actual = encoder.getDocumentXml();
-
-      if (expandedNodeId == null) {
-        // When encoding a null ExpandedNodeId the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "statusCode = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#statusCodeArguments")
-    void encodeStatusCode(@Nullable StatusCode statusCode, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeStatusCode("Test", statusCode);
-
-      String actual = encoder.getDocumentXml();
-
-      if (statusCode == null) {
-        // When encoding a null StatusCode the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "qualifiedName = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#qualifiedNameArguments")
-    void encodeQualifiedName(@Nullable QualifiedName qualifiedName, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeQualifiedName("Test", qualifiedName);
-
-      String actual = encoder.getDocumentXml();
-
-      if (qualifiedName == null) {
-        // When encoding a null QualifiedName the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "localizedText = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#localizedTextArguments")
-    void encodeLocalizedText(@Nullable LocalizedText localizedText, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeLocalizedText("Test", localizedText);
-
-      String actual = encoder.getDocumentXml();
-
-      if (localizedText == null) {
-        // When encoding a null LocalizedText the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "extensionObject = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#extensionObjectArguments")
-    void encodeExtensionObject(@Nullable ExtensionObject extensionObject, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeExtensionObject("Test", extensionObject);
-
-      String actual = encoder.getDocumentXml();
-
-      if (extensionObject == null) {
-        // When encoding a null ExtensionObject the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "dataValue = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#dataValueArguments")
-    void encodeDataValue(@Nullable DataValue dataValue, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDataValue("Test", dataValue);
-
-      String actual = encoder.getDocumentXml();
-
-      if (dataValue == null) {
-        // When encoding a null DataValue the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "variant = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.VariantArguments#variantOfScalarArguments")
-    void encodeVariantOfScalar(Variant variant, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeVariant("Test", variant);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "variant = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.VariantArguments#variantOfArrayArguments")
-    void encodeVariantOfArray(Variant variant, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeVariant("Test", variant);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "variant = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.VariantArguments#variantOfMatrixArguments")
-    void encodeVariantOfMatrix(Variant variant, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeVariant("Test", variant);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "diagnosticInfo = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ScalarArguments#diagnosticInfoArguments")
-    void encodeDiagnosticInfo(@Nullable DiagnosticInfo diagnosticInfo, String expected) {
-
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDiagnosticInfo("Test", diagnosticInfo);
-
-      String actual = encoder.getDocumentXml();
-
-      if (diagnosticInfo == null) {
-        // When encoding a null DiagnosticInfo the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-
-    @ParameterizedTest(name = "field = {0}, struct = {1}, dataTypeId = {2}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.StructArguments#structArguments")
-    void encodeStruct(String field, @Nullable UaStructuredType value, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      if (value != null) {
-        encoder.encodeStruct(field, value, value.getTypeId());
-      } else {
-        encoder.encodeStruct(field, null, NodeId.NULL_VALUE);
-      }
-
-      String actual = encoder.getDocumentXml();
-
-      if (value == null) {
-        // When encoding a null struct the encoder doesn't produce any XML
-        assertTrue(actual.isEmpty());
-      } else {
-        Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-        maybePrintXml(diff, expected, actual);
-
-        assertFalse(diff.hasDifferences(), diff.toString());
-      }
-    }
-  }
-
-  @Nested
-  class ArrayTests {
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#booleanArrayArguments")
-    void encodeBooleanArray(@Nullable Boolean[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeBooleanArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#sByteArrayArguments")
-    void encodeSByteArray(@Nullable Byte[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeSByteArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#byteArrayArguments")
-    void encodeByteArray(@Nullable UByte[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeByteArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#int16ArrayArguments")
-    void encodeInt16Array(@Nullable Short[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt16Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#uInt16ArrayArguments")
-    void encodeUInt16Array(@Nullable UShort[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt16Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#int32ArrayArguments")
-    void encodeInt32Array(@Nullable Integer[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt32Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#uInt32ArrayArguments")
-    void encodeUInt32Array(@Nullable UInteger[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt32Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#int64ArrayArguments")
-    void encodeInt64Array(@Nullable Long[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeInt64Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#uInt64ArrayArguments")
-    void encodeUInt64Array(@Nullable ULong[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeUInt64Array("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#floatArrayArguments")
-    void encodeFloatArray(@Nullable Float[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeFloatArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#doubleArrayArguments")
-    void encodeDoubleArray(@Nullable Double[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDoubleArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#stringArrayArguments")
-    void encodeStringArray(@Nullable String[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeStringArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#dateTimeArrayArguments")
-    void encodeDateTimeArray(@Nullable DateTime[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDateTimeArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#guidArrayArguments")
-    void encodeGuidArray(@Nullable UUID[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeGuidArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#byteStringArrayArguments")
-    void encodeByteStringArray(@Nullable ByteString[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeByteStringArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#xmlElementArrayArguments")
-    void encodeXmlElementArray(@Nullable XmlElement[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeXmlElementArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#nodeIdArrayArguments")
-    void encodeNodeIdArray(@Nullable NodeId[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeNodeIdArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#expandedNodeIdArrayArguments")
-    void encodeExpandedNodeIdArray(@Nullable ExpandedNodeId[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeExpandedNodeIdArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#statusCodeArrayArguments")
-    void encodeStatusCodeArray(@Nullable StatusCode[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeStatusCodeArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#qualifiedNameArrayArguments")
-    void encodeQualifiedNameArray(@Nullable QualifiedName[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeQualifiedNameArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#localizedTextArrayArguments")
-    void encodeLocalizedTextArray(@Nullable LocalizedText[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeLocalizedTextArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#extensionObjectArrayArguments")
-    void encodeExtensionObjectArray(@Nullable ExtensionObject[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeExtensionObjectArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#dataValueArrayArguments")
-    void encodeDataValueArray(@Nullable DataValue[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDataValueArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#variantArrayArguments")
-    void encodeVariantArray(@Nullable Variant[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeVariantArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-
-    @ParameterizedTest(name = "array = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.ArrayArguments#diagnosticInfoArrayArguments")
-    void encodeDiagnosticInfoArray(@Nullable DiagnosticInfo[] array, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeDiagnosticInfoArray("Test", array);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
-  }
-
-  @Nested
-  class MatrixTests {
-
-    @ParameterizedTest(name = "matrix = {0}")
-    @MethodSource(
-        "org.eclipse.milo.opcua.stack.core.encoding.xml.args.MatrixArguments#matrixOfBuiltinTypeArguments")
-    void encodeMatrixOfBuiltinType(@Nullable Matrix matrix, String expected) {
-      var encoder = new OpcUaXmlEncoder(context);
-      encoder.encodeMatrix("Test", matrix);
-
-      String actual = encoder.getDocumentXml();
-
-      Diff diff = DiffBuilder.compare(expected).withTest(actual).ignoreWhitespace().build();
-
-      maybePrintXml(diff, expected, actual);
-
-      assertFalse(diff.hasDifferences(), diff.toString());
-    }
+    assertFalse(diff.hasDifferences(), diff.toString());
   }
 
   /**
@@ -999,7 +67,7 @@ class OpcUaXmlEncoderTest {
    * @param expected the expected XML string.
    * @param actual the actual XML string.
    */
-  private static void maybePrintXml(Diff diff, String expected, String actual) {
+  static void maybePrintXml(Diff diff, String expected, String actual) {
     if (diff.hasDifferences() || DEBUG) {
       System.out.printf("Expected:%n%s%n", expected);
       System.out.printf("Actual:%n%s%n", actual);
